pyproject.toml
[project]
name = "fast-vc-service"
version = "0.1.5"
description = "Real-time voice conversion service based on Seed-VC, providing WebSocket voice conversion with PCM and Opus audio format support"
authors = [
    {name = "Leroll"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.10,<4.0"

dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.3",
    "loguru==0.7.3",
    "pydantic==2.11.5",
    "numpy==1.26.4",
    "resampy==0.4.3",
    "opuslib==3.0.1",
    "pyyaml==6.0.2",
    "torch==2.4.0",
    "torchvision==0.19.0",
    "torchaudio==2.4.0",
    "scipy==1.13.1",
    "librosa==0.10.2",
    "huggingface-hub==0.23.4",
    "transformers==4.46.3",
    "descript-audio-codec==1.0.0",
    "soundfile==0.12.1",
    "sounddevice==0.5.0",
    "funasr==1.1.5",
    "resemblyzer==0.1.4",
    "munch==4.0.0",
    "einops==0.8.0",
    "pydub==0.25.1",
    "jiwer==3.0.3",
    "gradio==4.44.0",
    "modelscope==1.18.1",
    "python-dotenv==1.1.0",
    "psutil==7.0.0",
    "fire==0.7.0",
    "onnxruntime-gpu>=1.22.0",
    "openpyxl>=3.1.5",
    "ipykernel==6.29.5",
    "pandas==2.3.0",
    "faiss-cpu>=1.12.0",
]


[project.scripts]
fast-vc = "fast_vc_service.cli:cli"

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
